moved localstorage check to main
